% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnreports-package.R, R/mnreports.R
\docType{package}
\name{mnreports-package}
\alias{mnreports}
\alias{mnreports-package}
\title{mnreports: State of Minnesota branded Markdown templates}
\description{
Includes MDH and MPCA state agency report themes with branded headers, chart themes, and tables. Produce reports in a variety of formats including HTML, MS Word, PDF, and Presentations. Use `mnreports::html_document` as the output format to create a smaller html file than the default `html_document`.
}
\seealso{
\link{html_document}, \link{pdf_document},
  \link{word_document},
}
\author{
\strong{Maintainer}: Dorian Kvale \email{dorian.l.kvale@state.mn.us}

Authors:
\itemize{
  \item Eric Kvale \email{eric.kvale@state.mn.us}
}

Other contributors:
\itemize{
  \item Chris Brueske [producer]
  \item Steve Gorg [producer]
}

}
\keyword{internal}
